+++ lib/eve_online/esi/models/allies.rb
@@ -5,7 +5,15 @@ module EveOnline
     module Models
       class Allies < Base
         def allies
+          output = []
 
+          return output if !options.is_a?(Array)
+
+          options.each do |ally|
+            output << Ally.new(ally)
+          end
+
+          output
         end
       end
     end